SystemUI: Separate FODAnimation into it's own class
